fix: fix missplace package for login and register
