+++ app/src/main/java/com/singaludra/gofood/feature/login/ui/LoginScreen.kt
@@ -1,4 +1,4 @@
-package com.singaludra.gofood.feature.register.ui
+package com.singaludra.gofood.feature.login.ui
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -17,7 +17,7 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.singaludra.gofood.feature.register.presentation.LoginUIState
+import com.singaludra.gofood.feature.login.presentation.LoginUIState
 import com.singaludra.gofood.shared.components.FilledButtonSection
 import com.singaludra.gofood.shared.components.InputPasswordTextFieldSection
 import com.singaludra.gofood.shared.components.OutlineButtonSection
