+++ app/src/main/java/com/singaludra/gofood/feature/register/ui/AddressScreen.kt
@@ -1,4 +1,4 @@
-package com.singaludra.gofood.feature.login.ui
+package com.singaludra.gofood.feature.register.ui
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -13,23 +13,22 @@ import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import com.singaludra.gofood.shared.components.FilledButtonSection
-import com.singaludra.gofood.shared.components.InputPasswordTextFieldSection
 import com.singaludra.gofood.shared.components.TextFieldSection
 import com.singaludra.gofood.shared.components.TopBarSection
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun RegisterScreen(
+fun AddressScreen(
     modifier: Modifier = Modifier,
+    dropDownList: List<String>,
     onArrowClick: () -> Unit,
-    onButtonClick: () -> Unit,
 ) {
     Scaffold(
         modifier = modifier,
         topBar = {
             TopBarSection(
-                title = "Register",
-                subTitle = "Register disini",
+                title = "Alamat",
+                subTitle = "Masukan Alamatmu",
                 onArrowClick = onArrowClick,
             )
         }
@@ -38,15 +37,25 @@ fun RegisterScreen(
             modifier = Modifier.padding(innerPadding),
             horizontalAlignment = Alignment.CenterHorizontally,
         ) {
-            Spacer(modifier = Modifier.weight(0.2f))
-            //add input photo section
 
             Spacer(modifier = Modifier.weight(0.2f))
 
-            Column(modifier = Modifier.padding(horizontal = 16.dp)) {
-               TextFieldSection(
-                    title = "Nama",
-                    placeholder = "Masukan Nama",
+            Column(
+                modifier = Modifier.padding(horizontal = 16.dp)
+            ) {
+                TextFieldSection(
+                    title = "No.Hp",
+                    placeholder = "Masukan nomor hp",
+                    keyboardOptions = KeyboardOptions.Default.copy(
+                        keyboardType = KeyboardType.Number,
+                        imeAction = ImeAction.Next
+                    ),
+                    onValueChange = {}
+                )
+                Spacer(modifier = Modifier.padding(vertical = 8.dp))
+                TextFieldSection(
+                    title = "Alamat lengkap",
+                    placeholder = "Alamat lengkapmu",
                     keyboardOptions = KeyboardOptions.Default.copy(
                         keyboardType = KeyboardType.Text,
                         imeAction = ImeAction.Next
@@ -55,32 +64,24 @@ fun RegisterScreen(
                 )
                 Spacer(modifier = Modifier.padding(vertical = 8.dp))
                 TextFieldSection(
-                    title = "Email",
-                    placeholder = "Masukan email",
+                    title = "Nomor rumah",
+                    placeholder = "Nomor rumahmu",
                     keyboardOptions = KeyboardOptions.Default.copy(
-                        keyboardType = KeyboardType.Email,
-                        imeAction = ImeAction.Next
-                    ),
-                    onValueChange = {}
-                )
-                Spacer(modifier = Modifier.padding(vertical = 8.dp))
-                InputPasswordTextFieldSection(
-                    title = "Password",
-                    placeholder = "Masukan Password",
-                    keyboardOptions = KeyboardOptions.Default.copy(
-                        keyboardType = KeyboardType.Password,
+                        keyboardType = KeyboardType.Number,
                         imeAction = ImeAction.Done
                     ),
                     onValueChange = {}
                 )
+                Spacer(modifier = Modifier.padding(vertical = 8.dp))
+
             }
 
             Spacer(modifier = Modifier.weight(1f))
 
             FilledButtonSection(
                 modifier = Modifier.padding(16.dp),
-                buttonText = "Lanjutkan",
-                onClick = onButtonClick,
+                buttonText = "Submit",
+                onClick = {}
             )
         }
     }
