+++ app/src/main/java/com/singaludra/gofood/feature/register/ui/RegisterScreen.kt
@@ -1,4 +1,4 @@
-package com.singaludra.gofood.feature.login.ui
+package com.singaludra.gofood.feature.register.ui
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -13,22 +13,23 @@ import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import com.singaludra.gofood.shared.components.FilledButtonSection
+import com.singaludra.gofood.shared.components.InputPasswordTextFieldSection
 import com.singaludra.gofood.shared.components.TextFieldSection
 import com.singaludra.gofood.shared.components.TopBarSection
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun AddressScreen(
+fun RegisterScreen(
     modifier: Modifier = Modifier,
-    dropDownList: List<String>,
     onArrowClick: () -> Unit,
+    onButtonClick: () -> Unit,
 ) {
     Scaffold(
         modifier = modifier,
         topBar = {
             TopBarSection(
-                title = "Alamat",
-                subTitle = "Masukan Alamatmu",
+                title = "Register",
+                subTitle = "Register disini",
                 onArrowClick = onArrowClick,
             )
         }
@@ -37,25 +38,15 @@ fun AddressScreen(
             modifier = Modifier.padding(innerPadding),
             horizontalAlignment = Alignment.CenterHorizontally,
         ) {
+            Spacer(modifier = Modifier.weight(0.2f))
+            //add input photo section
 
             Spacer(modifier = Modifier.weight(0.2f))
 
-            Column(
-                modifier = Modifier.padding(horizontal = 16.dp)
-            ) {
-                TextFieldSection(
-                    title = "No.Hp",
-                    placeholder = "Masukan nomor hp",
-                    keyboardOptions = KeyboardOptions.Default.copy(
-                        keyboardType = KeyboardType.Number,
-                        imeAction = ImeAction.Next
-                    ),
-                    onValueChange = {}
-                )
-                Spacer(modifier = Modifier.padding(vertical = 8.dp))
-                TextFieldSection(
-                    title = "Alamat lengkap",
-                    placeholder = "Alamat lengkapmu",
+            Column(modifier = Modifier.padding(horizontal = 16.dp)) {
+               TextFieldSection(
+                    title = "Nama",
+                    placeholder = "Masukan Nama",
                     keyboardOptions = KeyboardOptions.Default.copy(
                         keyboardType = KeyboardType.Text,
                         imeAction = ImeAction.Next
@@ -64,24 +55,32 @@ fun AddressScreen(
                 )
                 Spacer(modifier = Modifier.padding(vertical = 8.dp))
                 TextFieldSection(
-                    title = "Nomor rumah",
-                    placeholder = "Nomor rumahmu",
+                    title = "Email",
+                    placeholder = "Masukan email",
                     keyboardOptions = KeyboardOptions.Default.copy(
-                        keyboardType = KeyboardType.Number,
-                        imeAction = ImeAction.Done
+                        keyboardType = KeyboardType.Email,
+                        imeAction = ImeAction.Next
                     ),
                     onValueChange = {}
                 )
                 Spacer(modifier = Modifier.padding(vertical = 8.dp))
-
+                InputPasswordTextFieldSection(
+                    title = "Password",
+                    placeholder = "Masukan Password",
+                    keyboardOptions = KeyboardOptions.Default.copy(
+                        keyboardType = KeyboardType.Password,
+                        imeAction = ImeAction.Done
+                    ),
+                    onValueChange = {}
+                )
             }
 
             Spacer(modifier = Modifier.weight(1f))
 
             FilledButtonSection(
                 modifier = Modifier.padding(16.dp),
-                buttonText = "Submit",
-                onClick = {}
+                buttonText = "Lanjutkan",
+                onClick = onButtonClick,
             )
         }
     }
