+++ app/src/main/java/com/singaludra/gofood/main/factories/navigation/MainAppNavHost.kt
@@ -5,10 +5,10 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
+import com.singaludra.gofood.feature.login.ui.navigation.LOGIN_GRAPH_ROUTE
 import com.singaludra.gofood.feature.login.ui.navigation.REGISTER_GRAPH_ROUTE
-import com.singaludra.gofood.feature.login.ui.navigation.registerGraph
-import com.singaludra.gofood.feature.register.ui.navigation.LOGIN_GRAPH_ROUTE
-import com.singaludra.gofood.feature.register.ui.navigation.loginGraph
+import com.singaludra.gofood.feature.login.ui.navigation.loginGraph
+import com.singaludra.gofood.feature.register.ui.navigation.registerGraph
 
 @Composable
 fun MainAppNavHost(
